feat(front): add notebook page
