+++ javanotebook-front/src/App.tsx
@@ -1,63 +1,18 @@
 import * as React from 'react';
-import MonacoEditor from 'react-monaco-editor';
 import './App.css';
 
-import Button from '@material-ui/core/Button';
-import { API } from 'src/store/commands/api';
+import { NotebookPage } from './pages';
 
-interface IAppState {
-  code: string;
-}
-
-class App extends React.Component <{}, IAppState> {
+class App extends React.Component <{}, {}> {
 
 
-  public componentWillMount(){
-    this.setState({code: "//type your code here"})
-  }
-  
-  // constructor(props:any) {
-  //   super(props);
-    
-  // }
-
-  public handleSendCommand() {
-    API.sendCommand(this.state.code);
-    return;
-  }
 
   public render() {
-    const code = this.state.code;
-    const options = {
-      selectOnLineNumbers: true
-    };
     return (
       <div>
-        <MonacoEditor
-          width="800"
-          height="600"
-          language="java"
-          theme="vs-dark"
-          value={code}
-          options={options}
-          onChange={(value, e) => this.onChangeValue(value, e)}
-          editorDidMount={(editor) => this.editorDidMount(editor)}
-        />
-
-        <Button variant="contained" color="primary" className="App-ButtonSend" onClick={() => this.handleSendCommand()}>
-          Send
-        </Button>
+        <NotebookPage />
       </div>
     );
-  }
-  
-  public editorDidMount = (editor:any) => {
-    console.log("editor did mount");
-  }
-
-  public onChangeValue(newValue:any, e:any){
-    this.setState({code: newValue});
-    // console.log('onChange', newValue, e);
   }
 
 }
